requirement.txt
alembic==1.13.1
annotated-types==0.5.0
anyio==3.7.1
asttokens==2.4.1
bcrypt==4.1.2
certifi==2023.7.22
cffi==1.16.0
click==8.1.7
colorama==0.4.6
comm==0.2.0
cryptography==41.0.7
debugpy==1.8.0
decorator==5.1.1
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
executing==2.0.1
fastapi==0.103.1
greenlet==3.0.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
ipykernel==6.27.1
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.1
Mako==1.3.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
orjson==3.9.6
packaging==23.2
parso==0.8.3
passlib==1.7.4
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
psycopg==3.1.14
psycopg-binary==3.1.14
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyasn1==0.5.1
pycparser==2.21
pydantic==2.3.0
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.27.0
tornado==6.4
traitlets==5.14.0
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
uvicorn==0.23.2
watchfiles==0.20.0
wcwidth==0.2.12
websockets==11.0.3
